Revert "[WIP] add mingwx64"
