+++ app/src/main/java/dev/dimension/flare/ui/screen/compose/ComposeScreen.kt
@@ -96,6 +96,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.PopupProperties
 import androidx.navigation.NavBackStackEntry
+import app.cash.paging.compose.collectAsLazyPagingItems
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.isGranted
 import com.google.accompanist.permissions.rememberPermissionState
@@ -107,10 +108,13 @@ import com.ramcosta.composedestinations.navigation.DestinationsNavigator
 import com.ramcosta.composedestinations.spec.DestinationStyle
 import dev.dimension.flare.R
 import dev.dimension.flare.common.FileItem
+import dev.dimension.flare.common.collectAsState
 import dev.dimension.flare.data.datasource.bluesky.BlueskyDataSource
 import dev.dimension.flare.data.datasource.mastodon.MastodonDataSource
 import dev.dimension.flare.data.datasource.misskey.MisskeyDataSource
 import dev.dimension.flare.data.repository.ComposeUseCase
+import dev.dimension.flare.data.repository.accountServiceProvider
+import dev.dimension.flare.data.repository.activeAccountPresenter
 import dev.dimension.flare.model.MicroBlogKey
 import dev.dimension.flare.molecule.producePresenter
 import dev.dimension.flare.ui.component.NetworkImage
@@ -123,14 +127,11 @@ import dev.dimension.flare.ui.model.UiEmoji
 import dev.dimension.flare.ui.model.UiState
 import dev.dimension.flare.ui.model.UiStatus
 import dev.dimension.flare.ui.model.flatMap
+import dev.dimension.flare.ui.model.map
 import dev.dimension.flare.ui.model.onError
 import dev.dimension.flare.ui.model.onLoading
 import dev.dimension.flare.ui.model.onSuccess
-import dev.dimension.flare.ui.presenter.compose.ComposePresenter
-import dev.dimension.flare.ui.presenter.compose.ComposeStatus
-import dev.dimension.flare.ui.presenter.compose.MastodonVisibilityState
-import dev.dimension.flare.ui.presenter.compose.MisskeyVisibilityState
-import dev.dimension.flare.ui.presenter.compose.VisibilityState
+import dev.dimension.flare.ui.model.toUi
 import dev.dimension.flare.ui.theme.FlareTheme
 import dev.dimension.flare.ui.theme.screenHorizontalPadding
 import kotlinx.collections.immutable.toImmutableList
@@ -220,6 +221,18 @@ object ComposeTransitions : DestinationStyle.Animated {
             AnimatedContentTransitionScope.SlideDirection.Down,
         ) + fadeOut()
     }
+}
+
+private sealed interface ComposeStatus {
+    val statusKey: MicroBlogKey
+
+    data class Quote(
+        override val statusKey: MicroBlogKey,
+    ) : ComposeStatus
+
+    data class Reply(
+        override val statusKey: MicroBlogKey,
+    ) : ComposeStatus
 }
 
 @SuppressLint("MissingPermission")
@@ -346,7 +359,7 @@ private fun ComposeScreen(
                                     )
                                 }
                             }
-                            state.state.visibilityState.onSuccess { visibilityState ->
+                            state.visibilityState.onSuccess { visibilityState ->
                                 when (visibilityState) {
                                     is MastodonVisibilityState ->
                                         MastodonVisibilityContent(
@@ -371,7 +384,7 @@ private fun ComposeScreen(
                                     )
                                 }
                             }
-                            state.state.emojiState.onSuccess {
+                            state.emojiState.onSuccess {
                                 val isImeVisible = WindowInsets.isImeVisible
                                 IconButton(
                                     onClick = {
@@ -390,7 +403,7 @@ private fun ComposeScreen(
                             }
                         }
 
-                        state.state.emojiState.onSuccess { emojis ->
+                        state.emojiState.onSuccess { emojis ->
                             Box(
                                 modifier =
                                     Modifier
@@ -669,10 +682,10 @@ private fun ComposeScreen(
                     }
                 }
 
-                state.state.replyState?.let { replyState ->
+                state.replyState?.let { replyState ->
                     replyState.onSuccess { state ->
-                        if (state.itemCount > 0) {
-                            val item = state[0]
+                        if (state.listState.itemCount > 0) {
+                            val item = state.listState[0]
                             if (item != null) {
                                 UiStatusQuoted(
                                     status = item,
@@ -866,10 +879,12 @@ private fun composePresenter(
     status: ComposeStatus? = null,
     composeUseCase: ComposeUseCase = rememberInject(),
 ) = run {
-
-    val state = remember(status) {
-        ComposePresenter(status)
-    }.invoke()
+    val account by activeAccountPresenter()
+    val emojiState =
+        account.flatMap {
+            emojiPresenter(it).emojiState
+                ?: UiState.Error(IllegalStateException("Emoji not supported"))
+        }
     val textFieldState by remember {
         mutableStateOf(TextFieldState(""))
     }
@@ -877,15 +892,23 @@ private fun composePresenter(
         textFieldState.textAsFlow()
     }.collectAsState(initial = "")
     val pollState =
-        state.account.flatMap {
+        account.flatMap {
             when (it) {
                 is UiAccount.Bluesky -> UiState.Error(IllegalStateException("Bluesky not supported"))
                 is UiAccount.Mastodon, is UiAccount.Misskey -> UiState.Success(pollPresenter())
             }
         }
     val mediaState = mediaPresenter()
+    val visibilityState =
+        account.flatMap {
+            when (it) {
+                is UiAccount.Mastodon -> UiState.Success(mastodonVisibilityPresenter())
+                is UiAccount.Misskey -> UiState.Success(misskeyVisibilityPresenter())
+                is UiAccount.Bluesky -> UiState.Error(IllegalStateException("Bluesky not supported"))
+            }
+        }
     val contentWarningState =
-        state.account.flatMap {
+        account.flatMap {
             when (it) {
                 is UiAccount.Bluesky -> UiState.Error(IllegalStateException("Bluesky not supported"))
                 is UiAccount.Misskey, is UiAccount.Mastodon ->
@@ -894,10 +917,16 @@ private fun composePresenter(
                     )
             }
         }
-    state.replyState?.onSuccess {
-        LaunchedEffect(it.itemCount) {
-            if (it.itemCount == 1 && textFieldState.text.isEmpty()) {
-                when (val item = it[0]) {
+    val replyState =
+        status?.let { status ->
+            account.map {
+                statusPresenter(it, status)
+            }
+        }
+    replyState?.onSuccess {
+        LaunchedEffect(it.listState.itemCount) {
+            if (it.listState.itemCount == 1 && textFieldState.text.isEmpty()) {
+                when (val item = it.listState[0]) {
                     is UiStatus.Mastodon -> {
                         textFieldState.edit {
                             append("${item.user.handle} ")
@@ -935,8 +964,10 @@ private fun composePresenter(
         val canMedia = canMedia
         val pollState = pollState
         val mediaState = mediaState
+        val visibilityState = visibilityState
         val contentWarningState = contentWarningState
-        val state = state
+        val emojiState = emojiState
+        val replyState = replyState
 
         fun selectEmoji(emoji: UiEmoji) {
             textFieldState.edit {
@@ -946,7 +977,7 @@ private fun composePresenter(
 
         @RequiresPermission(Manifest.permission.POST_NOTIFICATIONS)
         fun send() {
-            state.account.onSuccess {
+            account.onSuccess {
                 val data =
                     when (it) {
                         is UiAccount.Mastodon ->
@@ -971,7 +1002,7 @@ private fun composePresenter(
                                     },
                                 sensitive = mediaState.isMediaSensitive,
                                 spoilerText = (contentWarningState as UiState.Success).data.textFieldState.text.toString(),
-                                visibility = ((state.visibilityState as UiState.Success<VisibilityState>).data as MastodonVisibilityState).visibility,
+                                visibility = (visibilityState as UiState.Success).data.visibility as UiStatus.Mastodon.Visibility,
                                 inReplyToID = (status as? ComposeStatus.Reply)?.statusKey?.id,
                                 account = it,
                             )
@@ -998,11 +1029,11 @@ private fun composePresenter(
                                     },
                                 sensitive = mediaState.isMediaSensitive,
                                 spoilerText = (contentWarningState as UiState.Success).data.textFieldState.text.toString(),
-                                visibility = ((state.visibilityState as UiState.Success<VisibilityState>).data as MisskeyVisibilityState).visibility,
+                                visibility = (visibilityState as UiState.Success).data.visibility as UiStatus.Misskey.Visibility,
                                 inReplyToID = (status as? ComposeStatus.Reply)?.statusKey?.id,
                                 renoteId = (status as? ComposeStatus.Quote)?.statusKey?.id,
                                 content = textFieldState.text.toString(),
-                                localOnly = ((state.visibilityState as UiState.Success<VisibilityState>).data as MisskeyVisibilityState).localOnly,
+                                localOnly = (visibilityState.data as MisskeyVisibilityState).localOnly,
                             )
 
                         is UiAccount.Bluesky ->
@@ -1023,6 +1054,39 @@ private fun composePresenter(
     }
 }
 
+@Composable
+private fun statusPresenter(
+    account: UiAccount,
+    status: ComposeStatus,
+) = run {
+    val service = accountServiceProvider(account = account)
+    val listState =
+        remember(account.accountKey) {
+            service.status(status.statusKey)
+        }.collectAsLazyPagingItems()
+
+    object {
+        val listState = listState
+    }
+}
+
+@Composable
+private fun emojiPresenter(account: UiAccount) =
+    run {
+        val service = accountServiceProvider(account = account)
+        val emojiState =
+            remember(account.accountKey) {
+                when (service) {
+                    is MastodonDataSource -> service.emoji()
+                    is MisskeyDataSource -> service.emoji()
+                    else -> null
+                }
+            }?.collectAsState()?.toUi()
+        object {
+            val emojiState = emojiState
+        }
+    }
+
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
 private fun contentWarningPresenter() =
@@ -1039,6 +1103,97 @@ private fun contentWarningPresenter() =
 
             fun toggle() {
                 enabled = !enabled
+            }
+        }
+    }
+
+internal sealed interface VisibilityState<T> {
+    val visibility: T
+    val showVisibilityMenu: Boolean
+    val allVisibilities: List<T>
+
+    fun setVisibility(value: T)
+
+    fun showVisibilityMenu()
+
+    fun hideVisibilityMenu()
+}
+
+internal abstract class MastodonVisibilityState(
+    override val visibility: UiStatus.Mastodon.Visibility,
+    override val showVisibilityMenu: Boolean,
+    override val allVisibilities: List<UiStatus.Mastodon.Visibility>,
+) : VisibilityState<UiStatus.Mastodon.Visibility>
+
+internal abstract class MisskeyVisibilityState(
+    override val visibility: UiStatus.Misskey.Visibility,
+    override val showVisibilityMenu: Boolean,
+    override val allVisibilities: List<UiStatus.Misskey.Visibility>,
+    val localOnly: Boolean,
+) : VisibilityState<UiStatus.Misskey.Visibility> {
+    abstract fun setLocalOnly(value: Boolean)
+}
+
+@Composable
+private fun misskeyVisibilityPresenter() =
+    run {
+        var localOnly by remember {
+            mutableStateOf(false)
+        }
+        var showVisibilityMenu by remember {
+            mutableStateOf(false)
+        }
+        var visibility by remember {
+            mutableStateOf(UiStatus.Misskey.Visibility.Public)
+        }
+        object : MisskeyVisibilityState(
+            visibility = visibility,
+            showVisibilityMenu = showVisibilityMenu,
+            allVisibilities = UiStatus.Misskey.Visibility.entries.toList(),
+            localOnly = localOnly,
+        ) {
+            override fun setLocalOnly(value: Boolean) {
+                localOnly = value
+            }
+
+            override fun setVisibility(value: UiStatus.Misskey.Visibility) {
+                visibility = value
+            }
+
+            override fun showVisibilityMenu() {
+                showVisibilityMenu = true
+            }
+
+            override fun hideVisibilityMenu() {
+                showVisibilityMenu = false
+            }
+        }
+    }
+
+@Composable
+private fun mastodonVisibilityPresenter() =
+    run {
+        var showVisibilityMenu by remember {
+            mutableStateOf(false)
+        }
+        var visibility by remember {
+            mutableStateOf(UiStatus.Mastodon.Visibility.Public)
+        }
+        object : MastodonVisibilityState(
+            visibility = visibility,
+            showVisibilityMenu = showVisibilityMenu,
+            allVisibilities = UiStatus.Mastodon.Visibility.entries.toList(),
+        ) {
+            override fun setVisibility(value: UiStatus.Mastodon.Visibility) {
+                visibility = value
+            }
+
+            override fun showVisibilityMenu() {
+                showVisibilityMenu = true
+            }
+
+            override fun hideVisibilityMenu() {
+                showVisibilityMenu = false
             }
         }
     }
