+++ shared/src/commonMain/kotlin/dev/dimension/flare/data/database/DriverFactory.kt
@@ -4,7 +4,7 @@ import app.cash.sqldelight.db.QueryResult
 import app.cash.sqldelight.db.SqlDriver
 import app.cash.sqldelight.db.SqlSchema
 
-expect class DriverFactory {
+internal expect class DriverFactory {
     fun createDriver(schema: SqlSchema<QueryResult.Value<Unit>>, name: String): SqlDriver
     fun deleteDatabase(name: String)
 }
