+++ shared/src/commonMain/kotlin/dev/dimension/flare/ui/presenter/status/StatusPresenter.kt
@@ -2,8 +2,7 @@ package dev.dimension.flare.ui.presenter.status
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import app.cash.paging.LoadState
-import app.cash.paging.LoadStateLoading
+import androidx.paging.LoadState
 import app.cash.paging.compose.LazyPagingItems
 import app.cash.paging.compose.collectAsLazyPagingItems
 import dev.dimension.flare.data.repository.activeAccountServicePresenter
@@ -26,7 +25,7 @@ class StatusPresenter(
         }
         val refreshing =
             listState is UiState.Loading ||
-                    listState is UiState.Success && listState.data.loadState.refresh is LoadStateLoading && listState.data.itemCount != 0
+                    listState is UiState.Success && listState.data.loadState.refresh is LoadState.Loading && listState.data.itemCount != 0
 
         return object : StatusState(
             listState,
