+++ shared/src/nativeMain/kotlin/dev/dimension/flare/KojectHelper.ios.kt
@@ -1,5 +1,8 @@
 package dev.dimension.flare
 
+import co.touchlab.kermit.Logger
+import co.touchlab.kermit.XcodeSeverityWriter
+import co.touchlab.kermit.platformLogWriter
 import com.moriatsushi.koject.Koject
 import com.moriatsushi.koject.start
